Changes to Game and a few others
Most new code is in Game... the function TakeTurn is a skeleton only.  I changed a few atributes of related Classes so it would show less error messages and created Class TicTacToe for the descrepencies that exist between game types. I used currentPlayer which is a class type instead of activePlayerID... If that's a bad idea, let me know.