Server and Client works... kinda
This solution might work for Fridays presentation. The Project works like before, with a twist.  Open up a Server console from the solution in folder LidgrenTest and then start a couple instances of the Monopoly Game. We need to figure out how to get the display to update without using the stupid button, but it seems to respond well enough.